Simulates two algorithms and compares them.
For neatness, the two different algorithms are defined in two methods demoAlgoOne(int turnCount) and demoAlgoTwo(int turnCount). Both accept an argument of the number of turns taken using that algorithm. These can easily be edited to fit whatever algorithm you want. The basis of the optimization AI is similar to these two function. The AI calls the function with more parameters, such as turn amount, what sum to reach before skipping, or whether or not to skip if a certain number if rolled. It runs trials of all possibilities and compares the outcomes. It has no sense of probability, just a desire to increase the average points-per-roll. AI coming soon... 

Note: The program can manage simulations with a huge number of trials, such as 3 million, but make sure to keep the number of turns low. In order to simulate "real life" better, a low number of turns is more accurate because human game play will most likely not exceed 100-200 turns per player. 

So far the largest simulation I have run is 10 million trials of 200 turns per player. 

Keep in mind that the period of Java's Random class is 2^48, so the dice rolls should be pretty random.